Add mockup API implementation from dev-team-docs
This is a placeholder, while we test the Continous Delivery system. We
need a project that looks real, and allows the API test suite to pass.
The mockup implementation from the API documentation works for this.

Once we start developing the system for real, the mockup will be
replaced.

Also include a copy of bottle.py, renamed to unifiedapibottle.py,
because we want version 0.12 of bottle.py, and Debian wheezy only
provides 0.10.

Also include a setup.py, which is needed for installation and Debian
packaging.