Fix bad call to resource's prepare_resource()
This bug was due to a half-done transition (rebased out of git
history) where prepare_resource used to be called with an argparse
arguments object. That worked, but it caused a stronger coupling
between a resource class and the application command line option
definitions. That coupling was broken by giving prepare_resource the
database URL directly, since that's (currently) the only thing that's
being used. If other things are needed by resources in the future,
then some other approach may need to be developed.

This bug was not found by the Continuous Delivery pipeline, because I
broke mine by trying to debug another problem and poking at my CD
directly, resulting in brokenness.

Found-by: Ville Niskala