Change the card resource field description into a table
This is a bit of an experiment to see what others think. I find the
table to be easier to read (both in source and when rendered), but more
troublesome to edit. I hope to get feedback on this in review, and if it
turns out to be unpopular, I'll change this back to a list.