Avoid using Database._conn directly
I've found a bug: selects use _conn directly, and this used to work in
sqlite, but doesn't in postgres. Part of this is that __init__ for
sqlite opens _conn, and it's valid for everything. Postgres does not.
This commit makes things more equal. It doesn't fix the bug, that's for
the future.