Add function table_name to construct table names
After this, unifiedapi.table_name is the ONLY allowed way to
construct names of database tables. ALL table names must be
done with this, even "simple" cases such as the table for the
resource type, which currently is just the name of the resource
type. If we ever want to change that table name, it's so much
easier to not have to vgrep through the code base to hunt down
all places where it's assumed.

Note that these names are not compatible with what we've done
previously. I don't care: this is part of a series of changes to
get rid of the need to empty databases on upgrades, and if we have
to do it one more time, so be it.

This drops the old Database.make_table_name method, which is now
useless, and fixes all call sites to use the new function. Except
in src/*, which will get a big change to how tables are created,
later on, so fixing those now is too much manual work.