Add Database.rename_table
When we update a resource tables, we'll do it by renaming the old
table, and creating a new one, and copying over the data. This is
simpler than modifying in-place, and based on Ville's anecodtal
evidence, it's also faster.

Also add a _in_transaction property to the base class, to be
overridden by subclasses, so that the shared rename_table method
can use it.