Separate storage prep from other startup
The storage preparation is now protected by a lock file. This is
because CREATE TABLE is NOT protected by transactions in PostgreSQL,
and, in fact, by the SQL standard. Also, things break if multiple
processes are doing it at the same time.

Using a lock file and only doing the prep when having the lock file is
the simplest way to make sure it only happens by one process at a time.