Change createtoken to get creds from config file
Instead of embedding the credentials in the createtoken source code, where
they'll end up in git and all sorts of leaking is possible, we'll have
createtoken get the credentials from a config file, kept outside of git.
Or at least not in the source code git, which is effectively public.

Also change test-api to tell createtoken, when run by yarn, where to
find the config file.