Rewrite eearch to use schemas, performance
This is quite perfect, yet. In fact, it's a kludge. However, it seems
to be a necessary kludge.

The old search implementation was using an ItemWalker, which works, but
results in quite complicated code. The new one uses schemas (lists of
tuples of table name, column name, and oolumn type), which are much
easier to deal with, and result in shorter, simpler code. This is good.

However, additionally, we have a performance problem, where we do a
simple SELECT on each table, and then use set intersections in Python to
get the desired result. It turns out that a lot of searches by API
clients find all the resources of a given type, or nearly all. For example,
a search might be for a "card_d_type" of "uid", which means all cards.
Such big results are very slow.

To speed that up, we need to do a more complicated SELECT. This turns out
to be a bit tricky, given our internal abstractions. Because the speedup
from the complicated SELECTs are needed urgently, due to ID06 imports
running in Sweden, we're doing a kludge, where we build the SELECT
and run it directly, bypassing a couple of layers of abstraction. The
abstractions will be fixed in the future, but we'll live without them now.

Additonally, the multi-table unit tests are disabled, because of differences
in SQL syntax between SQLite and PostgreSQL. They should be re-enable later,
when things are cleaned up. Yarn tests pass in any case.