Handle JSON-decoding of body explicitly
This is necessary because in Debian testing (stretch) bottle.py has,
for reasons unknown to me, started to sometimes use the simplejson
library instead of the json library included in the Python standard
library. The difference matters to us: simplejson returns plain
sgtrings, we need unicode strings.

By doing the parsing ourselves, we avoid this problem.