Change fake test PNGs be actual binary files
This is necessary to reproduce a bug in handling uploaded binaries.

Also, rename the test PNGs so it's easier to keep track of which is
which, when they're not all named the same.

We use the /usr/bin/printf explicitly, because the shell builtin
printf may not handle \xHH escapes.