Add slog fixes, unit tests for non-utf8 values
It turns out it's too late to convert values into utf8-safe text
during JSON encoding, because JSONEncoder doesn't allow us to override
the encoding for the types it knows about. Thus, we convert the values
as we get them into StructuredLogger. All strings are converted into
unicode, or else repr(). Lists, dicts, tuples are handled deeply,
meaning that any values inside them are also converted.

The types of values we convert are listed explicitly. Unknown types
cause an error via assert. This is because implicitly leaving such
values as-is results in hard-to-debug errors when JSONEncoder gets at
them. (The proof of this got rebased away.)