/any/ operator flag
/any/ operator flag mrings some sort of OR support for Qvarn search API.

For example, if you wan to search for foo=1 OR foo=2, you can do that with
/any/:

    /search/any/exact/foo/[1,2]

/any/ works with all vailable operators, you just need to add `/any/` before
operator and then  pass JSON encoded list of values to look for. All provided
values in list simple will be OR'ed.

As said above, /any/ works with all operators and also works with lists. For
example if `foo` field is a list, then all intersecting values will be checked.

Original proposal was to use _any suffix for all operators, like this:

    /search/exact_any/foo/[1,2]

But adding /any/ prefix was a bit easier to implement and is more consistend
with the rest of API.